% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz Function}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A real #}
}
\value{
looped 1:n with fizz buzz pattern
}
\description{
A Fizz Buzz function takes an integer as an argument.
}
\examples{
FizzBuzz(76)
FizzBuzz(3)

}
