% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather in Flagstaff, AZ}
\format{
a data frame with >25000 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{PRCP}{Amount of precipitation in inches.}
  \item{SNOW}{Amount of snow in inches.}
  \item{TMAX}{max of temp}
  \item{TMIN}{min of temp}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily maximum temperatures in Flagstaff, AZ.
}
\keyword{datasets}
